PasswordInput to separate module
